context: Fix implicit cast from ssize_t to int
We know the error value is contained between -4095 and 0, but MSVC still
complains that we might lose data with the implicit cast. Suppress the
warning with an explicit cast.

+++ context.c
@@ -626,7 +626,8 @@ iio_create_context_from_xml(const struct iio_context_params *params,
 		len = iio_snprintf(NULL, 0, "%s %s",
 				   ctx->description, description);
 		if (len < 0) {
-			prm_perror(params, -len, "Unable to set context description\n");
+			prm_perror(params, -(int) len,
+				   "Unable to set context description\n");
 			goto err_context_destroy;
 		}
